lógica com javaScript

+++ atividades/ex04b/ex04b.js
@@ -1,4 +1,4 @@
-let numero =parseInt(prompt("Digite um número"));
+let numero =parseFloat(prompt("Digite um número"));
 if(numero > 0 ){
     alert("Número é positivo");
 }else {
